% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{get_default_input}
\alias{get_default_input}
\title{Get default input for ACCEPT 3.0-CPRD}
\usage{
get_default_input()
}
\value{
A tibble with one default patient
}
\description{
Get default input for ACCEPT 3.0-CPRD
}
