% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{get_default_input}
\alias{get_default_input}
\title{Get default input for ACCEPT}
\usage{
get_default_input()
}
\value{
A data frame with one default patient
}
\description{
Get default input for ACCEPT
}
