% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{get_sample_input}
\alias{get_sample_input}
\title{Get sample input for ACCEPT}
\usage{
get_sample_input(n = NULL)
}
\arguments{
\item{n}{number of sample patients to return. Default returns all.}
}
\value{
A tibble of sample patients
}
\description{
Get sample input for ACCEPT
}
