% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{model_run}
\alias{model_run}
\title{Run ACCEPT 3.0-CPRD prediction model}
\usage{
model_run(model_input)
}
\arguments{
\item{model_input}{A tibble of patients in the same format as accept::samplePatients}
}
\value{
A tibble with predicted exacerbation probabilities and rates
}
\description{
Run ACCEPT 3.0-CPRD prediction model
}
